support --action_env vars in build scripts
The existing rustc_env attribute is useful when the provided env var
will not differ between machines, but is not practical when each
user of the build script may require a different value, such as passing
in a custom path to a tool on their machine.

Fixes the issue mentioned in
https://github.com/bazelbuild/rules_rust/pull/314#issuecomment-624524139
by starting from the default environment.