fail build on any warnings when running clippy
Currently warnings from rustc such as an unused import are printed
during first execution, but they do not fail the build. If cargo users
run clippy again, the same warnings get printed, but with Bazel nothing
is shown, as the previous success is cached.

This patch changes all warnings into errors, with the rationale being
that if we're failing for any clippy warning, it probably makes sense
to fail for rustc warnings as well.