rely on build script to locate bazel-out parent
Rust on Windows seems to struggle with relative pathnames in this case
- I suspect due to the fact that symlinks are being used.

fs::read_to_string("../../../path/to/somewhere")

gives an error

as does calling .canonicalize() on the above. Switching to \ as a
path separator does not seem to help.

If the parent directory references are handled separately, eg

Path::new("../../..").canonicalize(), then the path/to/somewhere
can be joined onto the folder and the file can be read. So instead
of automatically adjusting the path, we just return execroot verbatim,
and rely on the build script to do their own path manipulation.